requirements.txt
customtkinter
mysql-connector-python==8.0.33
python-dotenv
openai
Pillow